cambios del tipo de calcularCosto y cambio el calcularDias

+++ src/main/java/ar/edu/unq/poo2/envio/CorreoArgentino.java
@@ -1,5 +1,5 @@
 package ar.edu.unq.poo2.envio;
 
 public interface CorreoArgentino {
-	public  double estimarEnvio(double peso, Direccion destino);
+	public  float estimarEnvio(double peso, Direccion destino);
 }
